+++ backend/src/main/java/com/zwj/backend/service/CartService.java
@@ -10,13 +10,13 @@ public interface CartService {
     Result<List<CartItem>> getCart();
     
     // 添加商品到购物车
-    Result<CartItem> addToCart(Integer bookId, Integer quantity);
+    Result<CartItem> addToCart(Long bookId, Integer quantity);
     
     // 更新购物车项数量
-    Result<CartItem> updateCartItem(Integer cartItemId, Integer quantity);
+    Result<CartItem> updateCartItem(Long cartItemId, Integer quantity);
     
     // 删除购物车项
-    Result<Boolean> removeFromCart(Integer cartItemId);
+    Result<Boolean> removeFromCart(Long cartItemId);
     
     // 清空购物车
     Result<Boolean> clearCart();
@@ -25,8 +25,8 @@ public interface CartService {
     Result<Double> getCartTotal();
     
     // 通过ID查找购物车项
-    CartItem findCartItemById(Integer cartItemId);
+    CartItem findCartItemById(Long cartItemId);
     
     // 通过ID列表获取购物车项列表
-    List<CartItem> getCartItemsByIds(List<Integer> cartItemIds);
+    List<CartItem> getCartItemsByIds(List<Long> cartItemIds);
 }
