+++ backend/src/main/java/com/zwj/backend/service/OrderService.java
@@ -17,10 +17,10 @@ public interface OrderService {
     Result<Order> addOrder(OrderRequest request);
     
     // 在购物车项列表中查找指定ID的购物车项
-    CartItem findCartItemById(List<CartItem> cartItems, int cartItemId);
+    CartItem findCartItemById(List<CartItem> cartItems, Long cartItemId);
     
     // 删除订单
-    Result<Order> deleteOrder(int id);
+    Result<Order> deleteOrder(Long id);
     
     // 获取所有订单（管理员）
     Result<List<Order>> getAllOrders();
@@ -32,11 +32,11 @@ public interface OrderService {
     Result<Page<Order>> searchOrders(String keyword, LocalDateTime start, LocalDateTime end, int page, int pageSize);
     
     // 获取订单详情
-    Result<Order> getOrderDetail(int id);
+    Result<Order> getOrderDetail(Long id);
     
     // 支付订单
-    Result<Order> payOrder(int id, String paymentMethod);
+    Result<Order> payOrder(Long id, String paymentMethod);
     
     // 取消订单
-    Result<Order> cancelOrder(int id);
+    Result<Order> cancelOrder(Long id);
 }
